requirements.txt
scikit-ued
numpy
scipy
multiprocessing
crystals
pickle
ase
matplotlib
psutil
python-dotenv
